Fix - player model
- Player model의 teamId를 Team model의 id 값이 Long 타입으로 수정
- model 변경에 따른 code refactoring

+++ src/main/java/team/snowball/baseball/model/player/PlayerRepository.java
@@ -13,7 +13,7 @@ public interface PlayerRepository {
 
     int insert(Player player);
 
-    List<Player> findByTeamId(int id);
+    List<Player> findByTeamId(Long id);
 
     int delete(Long id);
